Updated build configurations in Eclipse for the runtime system.
There is a nasty bug in Eclipse: Before building the project, go into the project properties -> "C/C++ build" and just click "Apply" once. If this is not done, the build configuration is lost.

Also, the c:\cygwin\build directory needs to be in the path in order to run the debugger.

git-svn-id: https://svn.jmodelica.org@145 7ae51c65-13d5-4921-9d80-6511ddd1419c