Modified some files related to dfo in the optimization folder. Added input arguments "debug" and "plot_conv" in the function nelme and the function fmin in dfo.py.
debug is set to True to get separate error and output files for each separate process when using multiprocessing. plot_conv is set to True to get two plots with the convergence criteria for x and the objective function vs the iterations.

Added some more documentation in dfo.py. Also added the function galileo which is an OpenOpt function that works basically in the same way as de. (#1314)



git-svn-id: https://svn.jmodelica.org@2656 7ae51c65-13d5-4921-9d80-6511ddd1419c