#1494 jmodelica.optimization.polynomial:  * Merged RadauPol and RadauPol3 into RadauPol3.  * Created a new class called RadauPol, which represents a Lagrange polynomial of arbitrary degree used for Radau collocation.
CasadiRadau2 now supports an arbitrary amount of collocation points per element, using the new RadauPol class. Added tests for n_cp = 1, 3 and 8.

u_1_0 is currently not interpolated from succeeding values, and is thus ignored for all intents and purposes (which isn't as bad as it sounds). Will be fixed later.

Updated documentation and a few default values in algorithm_drivers.py.

git-svn-id: https://svn.jmodelica.org@2699 7ae51c65-13d5-4921-9d80-6511ddd1419c