Added a dfo example: a parameter estimation problem where a model of a Furuta pendulum is calibrated.
The optimization algorithm used is an implementation of the Nelder-Mead simplex algorithm which uses multiprocessing for the function evaluations. The file furuta_dfo_cost.py contains the definition of the objective function.

Note that when this example is run, sub-directories to the current one are created, where the function evaluations are performed. These must be manually removed after a run.

Two Python files, furuta_dfo.py and furuta_dfo_cost.py, were added in the folder jmodelica/examples. A mat file, FurutaData.mat, was added in jmodelica/examples/files and a fmu file, Furuta.fmu, was added in jmodelica/examples/files/FMUs.

git-svn-id: https://svn.jmodelica.org@2731 7ae51c65-13d5-4921-9d80-6511ddd1419c