Ticket #2024
Added two buttons to the toolbar (coolbar) for the documentationView. The buttons show a left and right arrow, symbolizing 'back' and 'forward' in the browser navigation history. The buttons only become visible when the active editor is the viewDocumentation editor. Furthermore, the buttons are inactive (greyed out) when there is no history in the corresponding direction.

The functionality is implemented using the command/handler pattern, and also uses the extensions "org.eclipse.ui.menus" for the menu contribution, as well as "org.eclipse.ui.services" for a sourceProvider (the NavigationProvider class) that provides global variables for accessing the navigation state.

git-svn-id: https://svn.jmodelica.org@3510 7ae51c65-13d5-4921-9d80-6511ddd1419c