A new parser target has been added, making it possible to parse strings into elements. This functionality is available in the ParserHandler. (#2037)
Formatting propagation now performs a check in all child nodes so that no new nodes with their own formatting has been added since the last propagation. If this is the case though, this formatting is now associated with corresponding nodes. PN_ComponentClauses and PN_FullClassDecls that are rewritten to ComponentDecls now also copy their formatting to the new nodes. These modifications are related to #2037, preparing for new nodes with their own formatting being added to the AST.

git-svn-id: https://svn.jmodelica.org@3527 7ae51c65-13d5-4921-9d80-6511ddd1419c