#2105 Added some functionality for this. Changes:  1. When a function call is made to a function that has non real outputs the referenced derivatives of the non real outputs, that are to be set by the function call, are set to null.  1. When a function call is made to a function that has non real inputs no derivatives for the non real inputs are passed in as arguments.  1. When declaring a function that has non real inputs no derivatives are declared in the header for the non real inputs.  1. For function calls with several outputs, function call equations, no derivatives of the non real variables are written back.  1.  Added a test.
I feel however that this might no be the optimal solution, especially for change 1 above. If there are variables with no derivatives present then the derivatives shouldn't be declared at all, instead of just setting them to null. Doing this requires changes in how the null check etc. is made in the function declarations, and I felt that it was best to leave it open for discussion for the time being. 


git-svn-id: https://svn.jmodelica.org@3658 7ae51c65-13d5-4921-9d80-6511ddd1419c