Ticket #2024 Added a "Generate documentation" option when right-clicking on a class in the project explorer. Before doing this, much of the code from the 'BrowserContent' class was factored out to a new class. All static HTML code (such as css) was put in this new class, as well as methods that given a (Full)ClassDeclaration generates HTML-code for a specific part (equation, imports, extensions etc.) of that class. The browserContent class is now only responsible to respond to user input, such as following links, maintaining browser history and handling javaScript input/output.
Documentation Generation:
1) When clicking on "Generate Documentation", the open editors are traversed to find out if the class is already open. If it is, this tab is brought to the front and if not, a new editor is opened for this class.

2) The user is requested to chose a directory from a JFileChooser. If this root directory already contains a subdirectory with the same name as the class (typically a library) the user is prompted to confirm an overwrite. Only directories are choosable.

3) Given a root directory 'c:/root', a library 'Lib' containing classes 'class1', 'class2', 'class3', each of which has an icon, the following directory structure is created ([d] denotes directory):

c:/root
	Lib [d]
		index.html
		icon.png
		Class1 [d]
			index.html
			icon.png
		Class2 [d]
			index.html (*)
			icon.png
		Class3 [d]
			index.html
			icon.png

The index.html files contains the HTML-code for the class with the same name as the directory they're in, e.g. index.html by (*) is the documentation for the class 'Lib.Class2'.

All file overwrites are automatic after the first prompt in 2).

Any hyperlink referring to a file contained in the directory or subdirectory to 'c:/root' will be clickable. Any other link, for example Modelica.Electrical.Analog.Basic.Ground will be rendered without a hyperlink. Whether or not a link to another class is clickable is determined by its full path; If it starts with 'Lib' its clickable, otherwise it's not. I'm not sure if this is the desired solution.

All javaScript has been stripped from the .html files stored to disk, as well as the 'edit' buttons for 'Information' and 'Revisions'.

git-svn-id: https://svn.jmodelica.org@3769 7ae51c65-13d5-4921-9d80-6511ddd1419c