Ticket #2024 Package wizard containing 3 classes added: org.eclipse.jface.wizard.Wizard handles the documentation generation. The wizard pops up after choosing "Generate documentation" by right-clicking on a class in the project explorer.
The first page, titled "Inclusions", asks "please specify what you want to include in the documentation", and has a series of SWT checkBoxes (Comment, Info, Import etc) with all of them pre-checked except "Equations" (as specified). Clicking next (which is always possible) takes the user to page 2.

The description of page 2 is: "Target" (the title) followed by "Please specify where you want to save the documentation". The content of the page is an empty SWT Text that and a button "browse" next to it. Back is enabled, next and finish grayed out. Entering text into the text field, or by choosing "browse" and selecting a folder from the JFileChooser, will enable the finish button. Clicking the finish button will create a pop-up containing a progress bar (org.eclipse.jface.dialogs.ProgressMonitorDialog). Above the progress indicator a text says "Generating documentation..." and below it a text that displays which class is currently being generated. There is a cancel button that if clicked cancels as soon as the current subTask (i.e class) is done.

Overriding existing files and cancelling the generation does not require confirmation by the user.

git-svn-id: https://svn.jmodelica.org@3817 7ae51c65-13d5-4921-9d80-6511ddd1419c