#3196 A public method for numerical evaluation of an expression is now added to the Model. This method calculates the values of the dependent parameters before the expression is evaluated. Another internal method can be used to evaluate MX expressions without recalculating the dependent parameters.
The internal method for numerical evaluation was changed, from substitute to creating a function on which numerical inputs are set (as this method was faster). 

Added a test and documentation. 

This method would benefit a lot if a way to keep track of changes to parameter values was introduced (also needed by #3178), as the dependent parameters would only be recalculated if needed. 

git-svn-id: https://svn.jmodelica.org@5436 7ae51c65-13d5-4921-9d80-6511ddd1419c