CasADiInterface: Implement dynamic typing of return values in the SWIG wrapper for #3171.
Added a method `void *_get_swig_p_type()` to all `SharedNode` subclasses that is included if creating a SWIG wrapper. The preprocessor define `-DMODELICACASADI_WITH_SWIG` controls this, and is always on right now. Each subclass should use the macro `MODELICACASADI_SHAREDNODE_CHILD_PUBLIC_DEFS` in its declaration to get this method.

Also had to change the format to call the SWIG macro `%instantiate_Ref` to be able to form the needed SWIG type names.


git-svn-id: https://svn.jmodelica.org@5677 7ae51c65-13d5-4921-9d80-6511ddd1419c