CasADiInterface: Implement forwarding of java exceptions to python for #3176.
Right now, java exceptions are printed to stdout, and then a Python exception is thrown with the text "a java error occurred; details were printed".
Exceptions thrown by JCC are caught in the same places as previously, but then rethrown.
All SWIG wrappers check for uncaught exceptions from JCC and report the error.

Also extend the try block in the transfer***Model functions to wrap the whole body.

git-svn-id: https://svn.jmodelica.org@5697 7ae51c65-13d5-4921-9d80-6511ddd1419c