Handle fixed delays by recording the delay data at the time it will be accessed, for #2620, cleaning up after r6908.
This allows time events to be triggered at the exact time when `update_position` will want to switch to the new interval.
If we want to reuse the same `jmi_delaybuffer_t` struct with different delay times, we will have to fix this issue in `update_position` instead.

Also use fix represent some boolean values using `TRUE` and `FALSE` instead of `1` and `0`.

git-svn-id: https://svn.jmodelica.org@6912 7ae51c65-13d5-4921-9d80-6511ddd1419c