Added runtime modules support.
Added the ability to modify the runtime behaviour of a functionality subset.
The desired behaviour is implemented by a static library and is linked with the
shared library in the FMU at compile time.

A new class, ModulesSettings, has been added which keeps track of the default
functionality subset modules. New runtime modules should register themselves
with the ModulesSettings class and provide the name of the static library
which implements the behaviour.



git-svn-id: https://svn.jmodelica.org@7752 7ae51c65-13d5-4921-9d80-6511ddd1419c