Merged revision(s) 9897 from trunk: #5225
Added a annotation provider based on a chained iterable, and changed annotations for short class declarations to use it. This in order to support the case where an annotation exists both for the declaration and a constrainedby clause.
........

Merged revision(s) 9904 from trunk:
#5368

Fixed so that DLLs from expanded libraries are included in an FMU.
........

Merged revision(s) 9934 from trunk:
Implemented ExtendsClause.name(). (#5383)
........



git-svn-id: https://svn.jmodelica.org@9942 7ae51c65-13d5-4921-9d80-6511ddd1419c